add seperate file for word validation unit tests with one additional test
